+++ examples/sps2019_with_frontends/main.tf
@@ -7,7 +7,7 @@ module "sharepoint" {
   sharepoint_version         = "2019"
   admin_username             = "yvand"
   number_additional_frontend = 2
-  add_public_ip_to_each_vm   = true
+  add_public_ip_address   = true
   rdp_traffic_allowed        = "10.20.30.40"
   enable_azure_bastion       = false
 }