+++ examples/spse_bastion_only_no_pip/main.tf
@@ -7,7 +7,7 @@ module "sharepoint" {
   sharepoint_version         = "Subscription-RTM"
   admin_username             = "yvand"
   admin_password             = "<admin_password>"
-  add_public_ip_to_each_vm   = false
+  add_public_ip_address   = false
   rdp_traffic_allowed        = "No"
   enable_azure_bastion       = true
 }