+++ main.tf
@@ -275,7 +275,7 @@ resource "azurerm_subnet_network_security_group_association" "nsg_subnetsp_assoc
 
 # Create artifacts for VM DC
 resource "azurerm_public_ip" "pip_dc" {
-  count               = var.add_public_ip_to_each_vm ? 1 : 0
+  count               = var.add_public_ip_address == "Yes" ? 1 : 0
   name                = "PublicIP-${local.config_dc["vmName"]}"
   location            = azurerm_resource_group.rg.location
   resource_group_name = azurerm_resource_group.rg.name
@@ -295,13 +295,13 @@ resource "azurerm_network_interface" "nic_dc_0" {
     subnet_id                     = azurerm_subnet.subnet_dc.id
     private_ip_address_allocation = "Static"
     private_ip_address            = local.network_settings["vmDCPrivateIPAddress"]
-    public_ip_address_id          = var.add_public_ip_to_each_vm ? azurerm_public_ip.pip_dc[0].id : null
+    public_ip_address_id          = var.add_public_ip_address == "Yes" ? azurerm_public_ip.pip_dc[0].id : null
   }
 }
 
 # Create artifacts for VM SQL
 resource "azurerm_public_ip" "pip_sql" {
-  count               = var.add_public_ip_to_each_vm ? 1 : 0
+  count               = var.add_public_ip_address == "Yes" ? 1 : 0
   name                = "PublicIP-${local.config_sql["vmName"]}"
   location            = azurerm_resource_group.rg.location
   resource_group_name = azurerm_resource_group.rg.name
@@ -320,13 +320,13 @@ resource "azurerm_network_interface" "nic_sql_0" {
     name                          = "ipconfig1"
     subnet_id                     = azurerm_subnet.subnet_sql.id
     private_ip_address_allocation = "Dynamic"
-    public_ip_address_id          = var.add_public_ip_to_each_vm ? azurerm_public_ip.pip_sql[0].id : null
+    public_ip_address_id          = var.add_public_ip_address == "Yes" ? azurerm_public_ip.pip_sql[0].id : null
   }
 }
 
 # Create artifacts for VM SP
 resource "azurerm_public_ip" "pip_sp" {
-  count               = var.add_public_ip_to_each_vm ? 1 : 0
+  count               = var.add_public_ip_address == "Yes" || var.add_public_ip_address == "SharePointVMsOnly" ? 1 : 0
   name                = "PublicIP-${local.config_sp["vmName"]}"
   location            = azurerm_resource_group.rg.location
   resource_group_name = azurerm_resource_group.rg.name
@@ -345,7 +345,7 @@ resource "azurerm_network_interface" "nic_sp_0" {
     name                          = "ipconfig1"
     subnet_id                     = azurerm_subnet.subnet_sp.id
     private_ip_address_allocation = "Dynamic"
-    public_ip_address_id          = var.add_public_ip_to_each_vm ? azurerm_public_ip.pip_sp[0].id : null
+    public_ip_address_id          = var.add_public_ip_address == "Yes" || var.add_public_ip_address == "SharePointVMsOnly" ? azurerm_public_ip.pip_sp[0].id : null
   }
 }
 
@@ -657,7 +657,7 @@ resource "azurerm_dev_test_global_vm_shutdown_schedule" "vm_sp_shutdown" {
 
 # Can create 0 to var.number_additional_frontend FE VMs
 resource "azurerm_public_ip" "pip_fe" {
-  count               = var.add_public_ip_to_each_vm ? var.number_additional_frontend : 0
+  count               = var.add_public_ip_address == "Yes" || var.add_public_ip_address == "SharePointVMsOnly" ? var.number_additional_frontend : 0
   name                = "PublicIP-${local.config_fe["vmName"]}-${count.index}"
   location            = azurerm_resource_group.rg.location
   resource_group_name = azurerm_resource_group.rg.name
@@ -677,7 +677,7 @@ resource "azurerm_network_interface" "nic_fe_0" {
     name                          = "ipconfig1"
     subnet_id                     = azurerm_subnet.subnet_sp.id
     private_ip_address_allocation = "Dynamic"
-    public_ip_address_id          = var.add_public_ip_to_each_vm ? element(azurerm_public_ip.pip_fe.*.id, count.index) : null
+    public_ip_address_id          = var.add_public_ip_address == "Yes" || var.add_public_ip_address == "SharePointVMsOnly" ? element(azurerm_public_ip.pip_fe.*.id, count.index) : null
   }
 }
 
@@ -739,7 +739,8 @@ resource "azurerm_virtual_machine_extension" "vm_fe_dsc" {
       "SQLName": "${local.config_sql["vmName"]}",
       "SQLAlias": "${local.general_settings["sqlAlias"]}",
       "SharePointVersion": "${var.sharepoint_version}",
-      "EnableAnalysis": false
+      "EnableAnalysis": false,
+      "SharePointBuildsDetails": ${jsonencode(local.sharepoint_subscription_packages)}
     },
     "privacy": {
       "dataCollection": "enable"